pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "ai-news-digest"
version = "0.1.0"
description = "Automatización mensual de noticias de IA: scraping, ranking por relevancia y armado de artículo listo para Medium."
readme = "README.md"
requires-python = ">=3.9"
authors = [{ name = "Tu Equipo D&A", email = "example@example.com" }]
license = { text = "MIT" }
dependencies = [
  "requests>=2.31.0",
  "beautifulsoup4>=4.12.3",
  "lxml>=5.2.2",
  "openai>=1.35.0",
  "lxml-html-clean>=0.1.1",   # <-- añade esto
  "pandas>=2.2.2",
  "click>=8.1.7",
  "tqdm>=4.66.4",
  "backoff>=2.2.1",
  "newsapi-python>=0.2.7",
  "newspaper3k>=0.2.8",
]


[project.scripts]
ai-news-digest = "ai_news_digest.cli:main"
